refactor: consolidate api-contracts, product specs, and backend utilities
- Move api-contracts/ to backend/api_contracts/ (backend infrastructure)
- Move product/ specs and prototypes into docs/product/ (consolidate docs)
- Move unused backend/app/ scripts to backend/scripts/:
  - fallback_game_manager.py, fix_wing_point_data.py
  - migrate_player_profiles.py, migrate_score_performance.py
- Consolidate conftest.py files: merge backend/ and backend/tests/ conftest
  into single coherent setup with path resolution and database init
- Remove redundant validate_deployments.py (superseded by verify_live_deployments.py)
- Update doc references for relocated api-contracts
