+++ backend/api_contracts/manager_contracts.py
