+++ backend/scripts/fallback_game_manager.py
