+++ backend/scripts/fix_wing_point_data.py
