+++ backend/scripts/migrate_player_profiles.py
