+++ backend/scripts/migrate_score_performance.py
