+++ docs/product/design-system/colors.json
