+++ docs/product/sections/live-scorekeeper/components/BettingActionsPanel.tsx
