+++ docs/product/sections/live-scorekeeper/components/NotesPanel.tsx
