+++ docs/product/sections/live-scorekeeper/components/PlayerCard.tsx
