+++ docs/product/sections/live-scorekeeper/components/StandingsSidebar.tsx
