+++ docs/product/sections/live-scorekeeper/components/TeamSelector.tsx
