+++ docs/product/sections/live-scorekeeper/components/index.ts
