+++ docs/product/sections/live-scorekeeper/spec.md
